+++ src/Components/Buttons/index.tsx
@@ -5,11 +5,10 @@ import { styles } from "./styles";
 const Buttons = ({title, onPress}) => {
     return(
         <TouchableOpacity
-         style={[styles.button]}
+        style={styles.button}
         onPress={onPress}>
             <Text style={styles.value}>{title}</Text>
         </TouchableOpacity>
     );
 }
-
 export default Buttons;
