+++ src/Components/CardMovies/index.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import { View, Text, Image, TouchableOpacity } from 'react-native';
+import { View, Image, TouchableOpacity } from 'react-native';
 import { styles } from './styles';
 import { MaterialIcons } from '@expo/vector-icons';
 import { Headline } from 'react-native-paper';
-
 interface CardProps {
   title: string;
   image: string;
@@ -20,7 +19,6 @@ export const Card: React.FC<CardProps> = ({ title, image, onPressAdd, onPressFav
           <Image style={styles.cardImage} source={{ uri: image }} />
           <View style={styles.titleContainer}>
             <Headline style={styles.cardTitle}>{title}</Headline>
-            
           </View>
         </TouchableOpacity>
 
@@ -29,7 +27,6 @@ export const Card: React.FC<CardProps> = ({ title, image, onPressAdd, onPressFav
             <TouchableOpacity onPress={onPressAdd}>         
             </TouchableOpacity>
           )}
-
           {onPressFavorite && (
             <TouchableOpacity style={styles.Icon} onPress={onPressFavorite}>
               <MaterialIcons style={styles.iconStyle} name={isFavorite ? 'favorite' : 'favorite-border'} size={24} color="red" />
