+++ src/Components/CardMovies/styles.ts
@@ -1,24 +1,19 @@
 import { StyleSheet } from "react-native";
 
 export const styles = StyleSheet.create({
-
   card:{
     justifyContent:"space-evenly"
   },
-
   cardContainerImage: {
     alignItems: "center",
     width: "100%",
-    
   },
-
   cardImage: {
     width: 160,
     height: 220,
     marginTop: 5,
     borderRadius: 10,
   },
-
   titleContainer: {
     fontWeight: 'bold',
     paddingHorizontal: 10,
@@ -31,23 +26,18 @@ export const styles = StyleSheet.create({
     height: 30,
     textAlign:"center"
   },
-
   iconsStyle: {
     justifyContent: "center",
     flexDirection: "row",
     alignContent:"flex-end"
   },
-
   Icon :{
-    
     justifyContent: "flex-start"
   },
-
   iconStyle: {
     marginRight: 15,
     padding: 10,
     alignItems:"flex-start",
     justifyContent: "flex-start"
   }
-
 })
