+++ src/Components/GitComponent/styles.ts
@@ -1,28 +1,26 @@
 import { StyleSheet } from "react-native";
 
 export const styles = StyleSheet.create({
-
-
-      gitComponent: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        marginBottom: 10,
-      },
-      cardImage: {
-        width: 50,
-        height: 50,
-        borderRadius: 25,
-        marginRight: 10,
-      },
-      textContainer: {
-        flexDirection: 'column',
-      },
-      name: {
-        fontSize: 16,
-        fontWeight: 'bold',
-      },
-      github: {
-        fontSize: 14,
-        marginTop: 1,
-      },
-    });
+  gitComponent: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginBottom: 10,
+  },
+  cardImage: {
+    width: 50,
+    height: 50,
+    borderRadius: 25,
+    marginRight: 10,
+  },
+  textContainer: {
+    flexDirection: 'column',
+  },
+  name: {
+    fontSize: 16,
+    fontWeight: 'bold',
+  },
+  github: {
+    fontSize: 14,
+    marginTop: 1,
+  },
+});
