--- src/Components/separador/index.tsx
@@ -1,1 +1,0 @@
-//Separador organizacional
