+++ src/screens/AboutUs/styles.ts
@@ -1,0 +1,55 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+
+  container: {
+    alignItems: "center"
+  },
+
+  title: {
+    fontSize: 30,
+    textAlign: 'center',
+    paddingTop: 30,
+    marginBottom: 5,
+  },
+
+  gitContainer: {
+    width: 280,
+    marginTop: 50
+  },
+
+  input: {
+    borderRadius: 5,
+    padding: 5,
+    marginBottom: 5,
+    borderWidth: 1,
+    width: 351,
+    height: 41
+  },
+
+  button: {
+    alignItems: "center",
+    backgroundColor: "#062451",
+    width: 351,
+    padding: 8,
+    borderRadius: 10,
+    marginTop: 2,
+
+  },
+
+  buttonTitle: {
+    color: 'white',
+  },
+
+  comment: {
+    backgroundColor: "#D9D9D9",
+    padding: 5,
+    margin: 2,
+    borderRadius: 5
+  },
+
+  scrollViewContent: {
+    paddingBottom: 160,
+
+  },
+})
