+++ src/screens/BemVindo/styles.ts
@@ -1,7 +1,6 @@
 import { StyleSheet } from "react-native";
 
 export const styles = StyleSheet.create({
-
     container: {
         flex: 1,
         backgroundColor: '#1CC8DF'
@@ -12,11 +11,11 @@ export const styles = StyleSheet.create({
         left: '20%',
     },
     title:{
+        textAlign: 'center',
         fontSize:30,
-        marginTop:370,
+        marginTop:500,
         color: "#fff",
         fontWeight:"500",
-        marginLeft:30
     },
     subTitle:{
         fontSize:15,
@@ -24,14 +23,10 @@ export const styles = StyleSheet.create({
         color: "#fff",
         fontWeight:"500",
         marginLeft:30
-
     },
-
     button: {
-        padding:12,
-        width: 300,
-        borderRadius:10,
-        marginTop: 100,
-        marginLeft:30
+        justifyContent: "center",
+        alignItems: "center",
+        margin: 150
     }
 });
