+++ src/screens/Favoritos/styles.ts
@@ -7,48 +7,34 @@ export const styles = StyleSheet.create({
     paddingHorizontal: 20,
     paddingTop: 10,
   },
-
   AreaView: {
   justifyContent: 'center',
   marginBottom: 135
   },
-
   titleText: {
     fontSize: 30,
     textAlign: 'center',
     paddingTop: 50,
     marginBottom: 5
   },
-
   noFavoritesContainer: {
     display: 'flex',
-    // justifyContent: 'center'
     alignContent: 'center',
     alignItems: 'center',
     marginTop: 100
   },
-
   noFavoritesText: {
     fontSize: 18,
     textAlign: 'center',
-    // fontWeight: 'bold',
-    
   },
-
   cardContainer: {
     borderRadius: 10,
-    // height: 420,
     width: 180,
     margin: 5,
     justifyContent: 'space-evenly',
-    // marginBottom: 10
-    
-
   },
-
   image: {
     height: 320,
     width: 250,
-    // marginLeft: 45
   }
 });
