+++ src/screens/Home/styles.ts
@@ -1,27 +1,23 @@
 import { StyleSheet } from "react-native";
 
 export const styles = StyleSheet.create({
-    
     container: {
         marginTop: 20,
         margin: 5
     },
-    
     title: {
         marginTop: 35,
         marginLeft: 8,
         fontWeight: 'bold',
         fontSize: 18
     },
-    scrollViewContent:{
+    scrollViewContent: {
         paddingBottom: 160,
     },
-
     showHideButtonText: {
-        marginLeft: 15 
+        marginLeft: 15
 
     },
-
     directorsFilter: {
         width: "100%",
         justifyContent: "space-evenly",
@@ -32,7 +28,6 @@ export const styles = StyleSheet.create({
         justifyContent: "space-evenly",
         width: "100%"
     },
-
     titleButtons: {
         backgroundColor: '#DDDDDD',
         marginRight: 10,
@@ -43,59 +38,48 @@ export const styles = StyleSheet.create({
         justifyContent: "center"
 
     },
-
     selectedDirector: {
         backgroundColor: 'black',
         color: 'white',
-      },
-
-   
-
+    },
     containerInput: {
         borderRadius: 16,
         height: 42,
         padding: 8,
         backgroundColor: "#67686D",
-        alignItems:'center',
+        alignItems: 'center',
         justifyContent: 'space-between',
-        flexDirection:"row"
+        flexDirection: "row"
     },
-
     headerText: {
         marginTop: 30,
         fontSize: 24,
         lineHeight: 45,
         fontWeight: 'bold'
     },
-
     input: {
-        height: 40, 
-        borderColor: 'gray', 
-        borderWidth: 1, 
-        margin: 10, 
+        height: 40,
+        borderColor: 'gray',
+        borderWidth: 1,
+        margin: 10,
         padding: 5,
-        backgroundColor:'#ffffffa3',
-        borderRadius:7
+        backgroundColor: '#ffffffa3',
+        borderRadius: 7
     },
-
     selectedButtonText: {
         color: '#fff',
-      },
-
-      buttonText: {
+    },
+    buttonText: {
         fontSize: 14,
-      },
-      selectedButton: {
+    },
+    selectedButton: {
         backgroundColor: '#2196F3',
-      },
-
-      cardContainer: {
+    },
+    cardContainer: {
         borderRadius: 10,
         width: 180,
         margin: 5,
         justifyContent: 'space-evenly',
 
-      },
-
- 
+    }
 })
