+++ src/services/Api.ts
@@ -14,7 +14,7 @@ export async function fetchFilms(): Promise<Film[]> {
   }
 }
 
-export async function fetchFilmById(id:string): Promise<Film[]> {
+export async function fetchFilmById(id: string): Promise<Film[]> {
   try {
     const response = await api.get(`/films/${id}`);
     return response.data;
